+++ src/commands/hardis/scratch/delete.ts
@@ -41,7 +41,7 @@ export default class ScratchDelete extends SfdxCommand {
   protected static requiresUsername = false;
 
   // Comment this out if your command does not support a hub org username
-  protected static supportsDevhubUsername = true;
+  protected static requiresDevhubUsername = true;
 
   /* jscpd:ignore-end */