+++ src/commands/hardis/work/new.ts
@@ -65,7 +65,7 @@ Under the hood, it can:
   protected static requiresUsername = false;
 
   // Comment this out if your command does not support a hub org username
-  protected static supportsDevhubUsername = true;
+  protected static requiresDevhubUsername = true;
 
   // Set this to true if your command requires a project workspace; 'requiresProject' is false by default
   protected static requiresProject = true;
@@ -163,7 +163,7 @@ Under the hood, it can:
           description: "Scratch orgs are configured on my project so I want to create or reuse one",
         },
         {
-          title: "Sandbox org with source tracking (beta)",
+          title: "Sandbox org with source tracking",
           value: "sandbox",
           description: "Release manager told me that I can work on Sandboxes on my project so let's use fresh dedicated one",
         },
@@ -303,5 +303,7 @@ Under the hood, it can:
         fail: true,
       });
     }
+    // Trigger a status refresh on VsCode WebSocket Client
+    WebSocketClient.sendMessage({ event: "refreshStatus" });
   }
 }