+++ src/commands/hardis/work/resetselection.ts
@@ -44,7 +44,7 @@ Calls a soft git reset behind the hood
   protected static requiresUsername = true;
 
   // Comment this out if your command does not support a hub org username
-  protected static supportsDevhubUsername = false;
+  protected static requiresDevhubUsername = false;
 
   // Set this to true if your command requires a project workspace; 'requiresProject' is false by default
   protected static requiresProject = true;